requirements.txt
pystray
psutil
pillow
pyinstaller